requirements.txt
tavily-python
python-dotenv
requests
